Fix implementation of GET_CURRENT_FRAME_ID on Linux
GET_CURRENT_FRAME_ID is used by the stack walker on Linux to figure out
the start of the frame. However, the implementation was still using the
win32 x64 implementation which use the return address as the frame
id. This broke the stack walker on Linux which is based on the x86
one. Changed the x64 linux implementation to match the x86 implementation.