Fixing an issue with symbol capturing in function defined in catch block's param
When catch block's param has a function definition, its node should be added to the list after the catch block. Right now we were adding the catch block to the list only after the params are parsed. Also during the visit the pnodescopes should be visited separate, not along with the params.