work for iteratorclose
This changeset enables IteratorClose scenario for destructuring, for..of
loop and built-ins (Promise, Array.from, Map, WeakMap, Set and WeakSet).
Closing the iterator means calling the return function of the iterator. The implementation for destructuring and for..of is done in the Bytecode level.
Destructuring : In the array destructuring pattern the try..finally
bytecodes are emitted in order to capture the abrupt completion of the
destructuring elements assignment.
For..of : The loop is surrounded by try..catch..finally. The catch block
is to tackle the abrupt completion due to exception and finally block is
to tackle the scenarios which involves break, return, continue.
Added unittest.