cross-platform: ChakraCore on OSX
Disclaimer: Currently, ChakraCore is not officially supported on OSX.

This PR targets the individuals interested with experimenting ChakraCore on
OSX. After seeing #1131, spent couple of hours to fix the compile, and link
time issues on OSX target.

Steps to compile :
```
brew install icu4c
./build.sh --debug -j:2 --icu=/usr/local/Cellar/icu4c/54.1/include/
```

Status:
CMAKE has an issue with shared bundles on OSX (a.k.a. so) and currently CH does not
play well with dylib.