Remove assert from OnBreakFlagChange which says it can only come in debug mode.
From the dump it looks like, edge was in sourcerundown mode and
during shutdown edge is deleting ScriptEngines and PDM is shutting down.
PDM is sending OnBreakFlagChange with APPBREAKFLAGS 0 on all the
engines in a loop. Since we already return failure if we are not
in debug mode I am removing the assert as this seems like it can
happen based on shutdown/delete sequence.