Merge Time-Travel Debugging
This PR adds the core time-travel debugging support code to ChakraCore and tests to ch.
See PR #1080 for comments -- earlier PR was closed + this PR opened to ensure naming/history is clean.

Tests are under test\TTBasic which test snapshot/restore behavior and test\TTExecuteBasic which test record/replay behavior.

The majority of the code is in lib\Runtime\Debug where 25 new files have been added. There are also a non-trivial of changes to the object/type representations (adding new virtual calls) and to the JsRT interface implementation. Minor other hooks have been added to other files as well (ScriptContext, DynamicObjectSnapshotEnumeratorWPCache, JavascriptExternalFunction, InterpreterStackFrame, etc.).

All changes are behind the ENABLE_TTD flag and this flag is only enabled on debug and test builds. When TTD is off the system requires a small number of fast checks on a tag on the ScriptContext so performance/memory impact is minimal when not enabled.