Precise closure-capture of local variables in the presence of deferred nested functions. This is a good thing to do anyway (improves CQ and solves some odd memory leak cases) and is especially important for redeferral and it allows consistency of byte code regardless of deferral. (Byte code will not necessarily be identical, but it will be functionally the same.) The parser is now responsible for detecting and marking non-local references to user variables, and we no longer force closure-capture of all locals in the presence of a deferred child function.
Precise closure-capture of local variables in the presence of deferred
nested functions. This is a good thing to do anyway (improves CQ and
solves some odd memory leak cases) and is especially important for
redeferral and it allows consistency of byte code regardless of
deferral. (Byte code will not necessarily be identical, but it will be
functionally the same.) The parser is now responsible for detecting and
marking non-local references to user variables, and we no longer force
closure-capture of all locals in the presence of a deferred child
function.