[MERGE #1188 @kunalspathak] improve changePrototype scenario
Merge pull request #1188 from kunalspathak:changeprototype

Whenever we change prototype of an object, we invalidate all proto inline caches of all the properties present in all of the object's prototype chain. However this can be avoided if the object is not a prototype of any other object since it would have never got registered in proto inline cache as a prototype object. So, skip invalidation if the object is not a prototype of any object.
The only thing that should be done in that case is to change the type of object so that if it's properties were using proto inline caches, they will be invalidated by type mismatch and we will update their proto inline cache on next property access.

Gives approx. 20% gain in AcmeAir benchmarks of node.

Test: UnitTest pass, Internal test runs in progress.