[MERGE #1233 @dilijev] Fix #36: class name in extends clause should be ReferenceError (use before declaration).
Merge pull request #1233 from dilijev:class

Previously would be TypeError or ReferenceError with a different message in many cases.

The spec changed at some point to make it so that the name of the class should effectively live in a temporary deadzone when the extends clause expression is being evaluated, but we were treating it as part of the scope in which the class statement or expression is being evaluated, which means that it could be undefined (in the case of `var x = (class x extends x`)) or not defined (in the case of `let x = (class x extends x)`), and similar behavior. In the case of `let x = ...` we would emit the correct error

`eval('x')` and other expressions using eval in the extends clause should also have the same semantics for binding `x` as long as we have the scopes organized correctly.

What this essentially meant was that we needed to move the parsing of the extends clause inside the class declaration's scope, and update the bytecode generator to reflect the new semantics.

Added a variety of tests to capture different cases to cover all the possibilities that produced different error cases. Also fixed up some tests that were no longer valid under the new semantics.

Thanks @ianwjhalliday for helping me understand the parser and bytecode semantics.
Thanks @bterlson for helping me understand the spec semantics.

Fixes #36.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/microsoft/chakracore/1233)
<!-- Reviewable:end -->