[MERGE #1202 @jordonwii] Implementation of Object.getOwnPropertyDescriptors.
Merge pull request #1202 from jordonwii:issue254_getownpropertydescriptors

Fixes #254

This is a rough implementation based on a combination of getOwnPropertyDescriptor as a reference, and what I could intuit from a day of digging.

I'd appreciate as much feedback as possible on how I did this. I used existing functions I was aware of when possible, but if there are better choices of helper functions here, please do let me know. I also did the tests all in javascript (similar to what I could find on getOwnPropertyDescriptor), and tried to make them as thorough as possible.