BUG 7944635: Replace calls to HiResTimer::Now() in ScriptSite::CallRootFunction with GetTickCount64().
Replace calls to HiResTimer::Now() in ScriptSite::CallRootFunction with
GetTickCount64(), which is a much less expensive alternative call, which also
has the additional advantage of being theoretically more accurate (it is not
affected by the various adjustments that can be made to system time while the
system is running).

The documentation of DateUtilities::TimeFromSt indicates:

// Returns number of milliseconds since Jan 1, 1970 [as a double]

In ScriptSite::CallRootFunction, we were using HiResTimer::Now() to get elapsed
time, so the fact that GetTickCount64() starts counting at system startup
instead of the epoch is unimportant.

We know that the units of DateUtilities::TimeFromSt are milliseconds
(represented as a double). GetTickCount64() also returns milliseconds (as an
unsigned long long, which is guaranteed by the standard to be at least 64 bits
on Windows and Linux in both x64 and x86), which we can convert to a double.

If we exceed 2^64 milliseconds on function time, we have bigger problems than
overflow of this timer (584 million years have elapsed).