[MERGE #1287 @jordonwii] Change JavascriptConversion::ToPropertyKey to have no return value.
Merge pull request #1287 from jordonwii:issue1251_voidtopropertykey

Fixes #1251

It turned out the *only* place the return value of ToPropertyKey was used (that I could find) was where I saw it in #1202. I changed the function signature to void, and removed the variable and the assert in that one place it was used.