[MERGE #1275 @Cellule] Memop bailout on negative length
Merge pull request #1275 from Cellule:memop

Bailout on Memset/Memcopy if the length is <= 0.
After the memop, we restore the induction variable to its expected value.
If the memop didn't actually happen we shouldn't modify the induction variable.
This is unlikely to happen since we would normally bailout on no profile and not emit this memop.