[MERGE #1298 @jordonwii] Add strong return types to some internal property enumerating functions
Merge pull request #1298 from jordonwii:issue1252

Fixes #1252

This adds strong return types (specifically, `JavascriptArray*`) to the following functions:

In JavascriptOperators:
1. `JavascriptOperators::GetOwnPropertyNames`
2. `JavascriptOperators::GetOwnPropertySymbols`
3. `JavascriptOperators::GetOwnPropertyKeys`
4. `JavascriptOperators::GetOwnEnumerablePropertyNames`
5. `JavascriptOperators::GetOwnEnumerablePropertyNamesSymbols`

The first 3 were mentioned in #1252, and 4 and 5 had essentially the same structure - all 5 just returned one of either JavascriptProxy::PropertyKeysTrap or one of the \*PropertiesHelper functions on JavascriptObject (listed below), so I went ahead and gave *all* of them a strong return type, as well.

In JavascriptObject (changed because the above all return the result of one of these):

1. `JavascriptObject::CreateOwnSymbolPropertiesHelper`
2. `JavascriptObject::CreateOwnStringPropertiesHelper`
3. `JavascriptObject::CreateOwnStringSymbolPropertiesHelper`
4. `JavascriptObject::CreateOwnEnumerableStringPropertiesHelper`
5. `JavascriptObject::CreateOwnEnumerableStringSymbolPropertiesHelper`
6. `JavascriptObject::CreateKeysHelper`

The first 5 all return the result of 6).

I had to update the signature of `JavascriptProxy::PropertyKeysTrap` to return `JavascriptArray*`, as well. This seemed correct, because a). most uses of it were followed by an assert that the result was a JavascriptArray, anyway, and b). [MDN says ](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Proxy/handler/ownKeys) the result of the ownKeys trap *must* be an array.

The rest of the changes are removing asserts and simplifying and/or removing any code that checked the type of the result of calls to any of the above functions.