[MERGE #1299 @xanlpz] runtests: set the PST timezone to run the tests
Merge pull request #1299 from xanlpz:timezone-tests

This reduces the failing tests from 21 to 1 if the right timezone is not set system-wide. The failing test comes from using PST and not Pacific Daylight Time, but not sure how to magically get the daylight stuff in here. In any case I think this is already an improvement.