[MERGE #1322 @satheeshravi] Fix for StrictMode with args optimization
Merge pull request #1322 from satheeshravi:StrictMode_BugFix

Dictionary type is not shareable with the current design and hence cannot be cached on the javascript library.
We can change this design if there is a compelling reason in the future.
For now, Creating dictionary type everytime still works better than what we had before.