Update the OOP JIT RPC interface bring up to support receiving security descriptors on both the server and the client.
This is necessary to be able to establish an RPC connection to a JIT process that is in another app container.

In addition we will not have access to the OpenProcess API for a process in another AppContainer, so we need to get a handle provided to us to be able to share the process handle across the RPC boundary.