[MERGE #1355 @pleath] Don't let parent function reference point to fake FunctionBody
Merge pull request #1355 from pleath:setnestedfunc

I've observed while investigating a class of site issues that the m_referenceInParentFunc field of the crashing FunctionBody in each case is pointing to the fake global FunctionBody shared by all deferred parses. I haven't been able to prove that this weirdness is responsible for the crashes, but it seems at the least like it's a good thing to fix.