[MERGE #1388 @pleath] Simple fix to tracking of nested functions containing 'with'
Merge pull request #1388 from pleath:with

Call to a nested function from within a 'with' requires that the function's enclosing scope be represented as an object. We were missing a case where the function is located in a param scope that will later be merge with a body scope. Also adding asserts to verify the object-ness of a scope when we're generating byte code that requires it.