[MERGE #1400 @curtisman] Cleanup JavascriptEnumerator interface
Merge pull request #1400 from curtisman:enumclean

The JavascriptEnumerator function GetCurrentAndMoveNext is misleading because.  Because it move to the next one before return the next element that we moved to. Change to MoveAndGetNext.

Also remove MoveNext/GetCurrentIndex/GetCurrentPropertyId, which is equivalent to GetCurrentAndMoveNext. Move all usage of those API to  MoveAndGetNext.