[MERGE #1447 @Cellule] Wasm: Compilation exception message bugfix
Merge pull request #1447 from Cellule:wasm/error

Fix bug in the Wasm Exception error message memory handling.
The memory for the string was being freed to soon by the destructor of the c++ exception. Which caused the next SysAllocString to allocate a different string at the same pointer of the previous wasm message.
The string's memory needs to be handled manually.