[PR #1168 @dilijev] Added console.log to jshost.exe and ch.exe as a convenient alias for WScript.Echo
Updated many test baselines that were affected by the addition of the console object to the global object.

[Core Commit Message]

[MERGE #1168 @dilijev] Added console.log to ch.exe as a convenient alias for WScript.Echo

Merge pull request #1168 from dilijev:chlog

This enables us to easily run examples written for the web browser that rely on console.log because they will now work in ch.exe without modification.

---

Short of merging this PR, here is a shim to work around this issue:

```js
if (typeof(console) === "undefined" &&
    typeof(WScript) != "undefined" &&
    typeof(WScript.Echo) != "undefined") {
    var console = {
        log: WScript.Echo
    }
}
```

If you include that in a test case (and you assume that you will never have a `console` object that doesn't have the `log` method defined, and you only use the `WScript.Echo` and `console.log` methods in mutually compatible ways) you can reuse the entire text of the test between browser hosts and the `ch.exe` host.