[MERGE #1482 @ThomsonTan] Fix trashing scratch register (r12) by the call of chkstk in prolog code
Merge pull request #1482 from ThomsonTan:FixR12TrashingOnArm

When function or loopbody needs chkstk, we utilize scratch register (r12) to call it in prolog. At the end of prolog, we assume scratch register still holds zero and use it to initialize CallInfo in InlineedFrame, then we get corrupted InlineedFrame on the stack. See below snnippet in prolog for details:

$L34123:
    (r12).i32       =  MOV            0 (0x0).i32
$L34124:
    [(sp).i32].i32  =  PUSH           0, 1, 2.i32
    [(sp).i32].i32  =  PUSH           11, 14.i32
    (r11).i32       =  MOV            (sp).i32
    [(sp).i32].i32  =  PUSH           4, 5, 6, 7, 8, 9, 10, 12.i32
    [(sp).i32].i32  =  VPUSH          8, 9, 10, 11, 12, 13, 14, 15.i32
    (r4).i32        =  LDIMM          4037 (0xFC5).i32
    (r12).i32       =  LDIMM          CRT_chkstk.u32
    (r4).i32        =  BLX            (r12).i32
    (sp).i32        =  SUB            (sp).i32, (r4).i32
    iarg65535(s94418)<0>.i32 = STR    (r12).i32