Medusa Branch Mirrorer (replaces VSO->GitHub Pump)
This is a replacement for the old VSO->GitHub pump. That pump was meant for short term use for one branch, however we ended up relying on it and it was in dire need of a rewrite.

This new mirroring service takes a set of remotes and branch mappings for those remotes. It works out which branches need mirroring and fast forwards the destinations to match the source. If there are incoming commits, it will email the branch owners and track when the branch has been restored. It is designed to be independent and portable.

Improvements:
- No more reliance on Razzle perl. Please have the dependencies installed listed at the beginning of the file. (For Strawberry Perl, you would use 'cpan Capture::Tiny Mail::Sender' for example.)
- We check for branch differences before attempting to mirror.
- Remotes are fetched first to speed up branch checks.
- Email sending no longer relies on Razzle perl libs.
- A blacklist of bad branches is maintained. Branches on the blacklist will periodically send reminder emails at a specified interval until the problem has been resolved.
- Removed dependence on Sync script libs.

Testing:
- It has been the branches in this config for over a week.
- Manually tested conflicting commits using a branch on GitHub.