[MERGE #1491 @Cellule] QuickSort
Merge pull request #1491 from Cellule:quicksort

Remove our own implementation of QuickSort in favor of the stdlib version.

This implementation was used by SwitchBuilder which could run out of stack if we had a very long list of cases already sorted.
It is hard to hit in the release product has compiler optimization changes the number of cases needed to hit the issue and we don't jit functions past a certain bytecode size.
I was able to repro the issue with a switch of 16000 cases in debug.
This problem is more problematic for WebAssembly as we will jit very huge functions no matter what and cases are always sorted