[MERGE #1488 @meg-gupta] Fix for cleanup of empty functions in presence of -bailoutoneveryline and related flags
Merge pull request #1488 from meg-gupta:emptyfix

Flag -bailoutateveryline injects bailout code at every line. On a bailout, we assert if there is an argument object used in the function body, it should be either zero initialized or valid. We eliminate zero initialization of arg object when stack args is enabled because it is initialized again. But in the presence of bailout injection flag, we bailout much before the argument object is initialized. So in the presence of this flag, we should not eliminate zero initialization.