[MERGE #1555 @MikeHolman] fix issue where passing global object fails asm.js validation
Merge pull request #1555 from MikeHolman:simdlinkissue

We were doing lookup to see if the library object that the user passed has SIMD, and then we look at the SIMD properties which are used. So even if SIMD is disabled and unused, we still do that one property lookup of SIMD itself. Since SIMD is not on the global object (this is behind a flag), the property lookup fails and we need to go to the host to complete the lookup. Unfortunately, when this happens in Edge, we mark the property lookup as potentially having side effects and abort asm.js validation (as side effecting operations may cause observable semantic differences when running in js vs. asm.js).

This change prevents doing a property lookup while linking unless the property is actually used.