[MERGE #1590 @dilijev] Fix some tabs and trailing whitespace that has creeped in; add Tab Check
Merge pull request #1590 from dilijev:whitespace

Add a new Jenkins task to detect when tabs are introduced into the source code.

There are very few instances of tabs being introduced (always unintentionally in the case of C++ source) and it seems to bother enough people that it would be nice to detect it before it is checked in.