[MERGE #1650 @MikeHolman] fix xdata issues on arm
Merge pull request #1650 from MikeHolman:thumbencoding

We were masking the code address for thumb mode before registering the unwind data with the OS, which causes unwinding to fail.

Also, for OOP JIT specifically, when we have expanded unwind data we need to fixup the .xdata pointer. Also fixed an issue I recently introduced which was causing bad alignment for everything in NativeDataBuffer (necessary here as .xdata is required to be 4 byte aligned).