[MERGE #1646 @akroshg] Passing proper property operation flag during delete operation.
Merge pull request #1646 from akroshg:del1

On many Array methods, as per spec, we need to throw when we try to delete
the non-configurable property. Earlier we were passing wrong flag
PropertyOperation_ThrowIfNotExtensible. Fixed that by passing the correct
flag and throw when that happens.
Added few tests for reference.