[MERGE #1579 @rajatd] Regex backtracking improvements
Merge pull request #1579 from rajatd:regex

This PR improves the algorithm for backtracking through a set with a * or a + wildcard.

When we get an unsuccessul match after such a set, we used to backtrack one character at a time over the matched characters of the set. The optimization being made in this change is that if we know at compile time what needs to follow the set for a successful match, we could record the offsets of that (those) character(s) while matching the set, and then backtrack only to those offsets.