[MERGE #1675 @MikeHolman] fix bug with trying to use full JS strings from JIT
Merge pull request #1675 from MikeHolman:oopstring

For OOP JIT, we only marshal across the necessary bits we need from RecyclableObjects, in particular for JavascriptStrings we only marshal the typeId, buffer, and length.
We then need to make sure to only use the bits we have available.

There was a place where we do FromVar, which under an assert calls type->CanHaveInterceptors(). And that is data we don't have, so it reads some invalid memory from the RPC buffer.

To make this all much safer (and fix the assert) I have added a new JITJavascriptString class which contains only the fields we have, and replaced all backend string usage with this. This also gave me the opportunity to enable a couple of places where we were being cautious and disabling string optimizations for OOP JIT (these optimizations were pretty minor, so OOP JIT benchmarks are flat).