[MERGE #1695 @agarwal-sandeep] Set debugContext to null before calling close and HeapDelete
Merge pull request #1695 from agarwal-sandeep:debugcontextdelete

In multiple context scenarios (Slate) with chk builds I see debugContext
is not null and is garbage, ScriptContext is in closed state which means
debugContext is in the process of being deleted and has not be set to null.
I can put critical section to the particular AV but it’s better to hold
debugContext in a temporary variable set scriptContext->debugContext to
null and then HeapDelete memory pointed by temporary variable.

I have kept the autoDebugContextCloseCS for PDM case though it should
not be needed (I think so)