[MERGE #1681 @Penguinwizzard] If we have no user crash handler we should generate additional info.
Merge pull request #1681 from Penguinwizzard:bug9006509

The previous way we were doing things, we'd only generate the additional
backtrace information about arguments and the like when an exception was
_totally_ uncaught. This interfered with our implicit autocatchhandlers,
which were added for some instructions. This caused divergences with the
nonative version of the debug output, where we wouldn't add these, which
caused us to only generate the additional backtrace information in a few
of the test configurations, but not all of them. By allowing creation of
the additional backtrace information when we have a catch handler but no
user catch handler, we handle these cases more uniformly.

This does, however, mean that our information on whether or not there is
a user catch handler; the existing code does this to some degree, but it
is polluted by a bunch of implied catch handlers in both interpreter and
in helper code. I added an argument to autocatchhandler to specify if we
can mark a particular instance of the autocatchhandler as user-made; the
values that I've chosen for it (i.e. where it is true) should be checked
and verified. Note that if the argument is true, that only means that it
_can_ be a user catch handler; if it's not in user code (e.g. in chakra-
generated JS) then it still won't be marked as a user catch handler.