[MERGE #1709 @meg-gupta] Fix for AddrOpnd to IntConstOpnd tests break
Merge pull request #1709 from meg-gupta:immbug

When a IntConstOpnd fits in the opnd field, but the instruction doesnt
accept an int constant as its operand, then the immediate should be
hoisted into a register and then used in the instruction. Fix the conditions so
that this case is handled correctly.