[MERGE #1673 @Penguinwizzard] Fix issue with unsigned compare peep in dead code.
Merge pull request #1673 from Penguinwizzard:fg_fix

The unsigned compare peep optimization, after being changed to have
the bytecodeuses instruction emitted after the compare instead of before
it, had issues with lowering of code hidden behind a bailonnoprofile, in
that it would have conflicts on lifetimes due to the loss of atomicity
of the bytecodereg use and generation of the compare. This patch attempts
to restore that by aggregating the bytecodeuse information for a single
instruction into once place.