[MERGE #1725 @Cellule] ByteCodeWriter offset reasoning bug
Merge pull request #1725 from Cellule:users/micfer/opcode

Revert part of the extended opcode change #1444. There was an error in reasoning about bytecode offset before EncodeT.
I thought the next thing to be encoded was always the opcode, this is not true if there is a branch Island created there.